sql: create `sql/sessiondatautil` package
This commit is to move `sql.NewFakeSessionData()` function to a new package
`sql/sessiondatautil`.

Release note: none
