+++ pkg/BUILD.bazel
@@ -1564,6 +1564,7 @@ GO_TARGETS = [
     "//pkg/sql/sessiondata:sessiondata",
     "//pkg/sql/sessiondata:sessiondata_test",
     "//pkg/sql/sessiondatapb:sessiondatapb",
+    "//pkg/sql/sessiondatautil:sessiondatautil",
     "//pkg/sql/sessioninit:sessioninit",
     "//pkg/sql/sessioninit:sessioninit_test",
     "//pkg/sql/sessionphase:sessionphase",
@@ -2540,6 +2541,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/sql/sem/volatility:get_x_data",
     "//pkg/sql/sessiondata:get_x_data",
     "//pkg/sql/sessiondatapb:get_x_data",
+    "//pkg/sql/sessiondatautil:get_x_data",
     "//pkg/sql/sessioninit:get_x_data",
     "//pkg/sql/sessionphase:get_x_data",
     "//pkg/sql/span:get_x_data",
