+++ pkg/server/BUILD.bazel
@@ -194,6 +194,7 @@ go_library(
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
         "//pkg/sql/sessiondatapb",
+        "//pkg/sql/sessiondatautil",
         "//pkg/sql/sessioninit",
         "//pkg/sql/sqlinstance",
         "//pkg/sql/sqlinstance/instanceprovider",
@@ -416,6 +417,7 @@ go_test(
         "//pkg/sql/sem/catconstants",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
+        "//pkg/sql/sessiondatautil",
         "//pkg/sql/sqlstats",
         "//pkg/sql/sqlstats/persistedsqlstats",
         "//pkg/sql/tests",
